Perform logical checks on a vector of institutions using ChatGPT.

For example, we might be interested in checking whether author institutions are a nonprofit, private company, or university.

Usage

ll_find_institution_type(
  text,
  model = "gpt-4o-mini",
  api_key = Sys.getenv("OPENAI_API_KEY"),
  logical_check = "in the United States"
)

Arguments

text

A vector of strings containing cleaned institutions. These might be the output of `ll_pull_institutions`.

model

string, which OpenAI ChatGPT model to use.

api_key

string, the API key if you want to supply it manually. There is no free one, but used with a minimal model is very cheap. https://openai.com/api/pricing/.

logical_check

string, a phrase that can be answered "Yes" or "no" when combined with "Is this institution...". For example, "in the United States" or "a Minority Serving Institution".

Value

A data frame containing T/F data for each item in `text`.

Examples

ll_find_institution_type("Johns Hopkins University", logical_check = "an institution with >5000 students")
#> No encoding supplied: defaulting to UTF-8.
#> Error in content$choices: $ operator is invalid for atomic vectors